osaibot.c: Move all run time communication with socket into seperate pseudo-threads
The main bath process will no longer see the socket at runtime, so
the path to dup the fd into somewhere writable by an arbitrary process
and mess with it, this path is now closed.

To get a generic clone (with non-shared fd table) with pthread_create
I can only describe as pain. Could call clone ourselves but then we'd
need to implement malloc and locks. Could do disassembly and rewrite
syscall instruction with int3 but the complexity... I settled with
seccomp to generate a SIGSYS but that's still a can of worms.